backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
psycopg2-binary
requests
beautifulsoup4
langchain
langchain-google-genai
python-multipart
python-dotenv
